requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.12.0
djangorestframework==3.15.2
mysqlclient==2.2.7
pip==25.0
python-dotenv==1.0.1
setuptools==75.8.0
sqlparse==0.5.2
typing_extensions==4.12.2
wheel==0.45.1
